src/requirements.txt
canvasapi==3.0.0
psycopg3
